% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderated t-covariate functions.R
\name{create.t.stat.cov.df}
\alias{create.t.stat.cov.df}
\title{Create moderated t-statistics and covariates}
\usage{
create.t.stat.cov.df(Y, X, cont, lmFit.args = list(), eBayes.args = list())
}
\arguments{
\item{Y}{matrix containing log-ratios or log-expression values for a series of arrays, with rows corresponding to genes and columns to samples.}

\item{X}{design matrix of the microarray experiment, with rows corresponding to samples and columns to coefficients to be estimated. The matrix must have at least 3 columns and the rows must sum to 1.}

\item{cont}{contrast numeric for the intended hypothesis test.}

\item{lmFit.args}{list of further arguments for the lmFit function from the limma package.}

\item{eBayes.args}{list of further arguments for the eBayes function from the limma package.}
}
\value{
A list with the elements
\item{mod.t.stat}{numeric of moderated t-statistics.}
\item{mod.t.cov}{numeric of moderated t-covariates.}
\item{df}{numeric of degrees of freedoms for the null component.}
}
\description{
Create moderated t-statistics and covariates
}
\examples{
create.t.stat.cov.df(exampleY, exampleX, c(1,-1,0,0,0))
}
