% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM tmm.R
\name{fit.t.model}
\alias{fit.t.model}
\title{Fit a neural network t-mixture model}
\usage{
fit.t.model(
  mod.t.cov.df,
  tol = 1e-07,
  alt.df = NULL,
  n.iter = NULL,
  max.iter = 25,
  n.nodes = NULL
)
}
\arguments{
\item{mod.t.cov.df}{list returned from \link[modtcov]{create.t.stat.cov.df}.}

\item{tol}{tolerance control for the "L-BFGS-B" method.}

\item{alt.df}{numeric of the degrees of freedom for the non-null t-mixture model components. Taken to be the same as the null component by default.}

\item{n.iter}{number of iterations for the EM algorithm.}

\item{max.iter}{maximum number of iterations for the EM algorithm.}

\item{n.nodes}{number of nodes in the hidden layer of the neural network.}
}
\value{
A list with the elements
\item{param}{numeric of fitted mu parameters and 1/tau parameters.}
\item{prob_mat}{numeric of fitted conditional probabilities.}
\item{nnobj}{object from \link[nnet]{nnet}.}
\item{mod.t.stat}{numeric of moderated t-statistics.}
\item{mod.t.cov}{numeric of moderated t-covariates.}
\item{df}{numeric of degrees of freedoms for the null component.}
\item{alt.df}{numeric of degrees of freedoms for the non-null components.}
}
\description{
Fit a neural network t-mixture model
}
\examples{
mod.t.cov.df <- create.t.stat.cov.df(exampleY, exampleX, c(1,-1,0,0,0))
fit <- fit.t.model(mod.t.cov.df)
}
