% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlfdr tmm.R
\name{mlfdr_fn_tmm}
\alias{mlfdr_fn_tmm}
\title{Create mimicked local false discovery rates}
\usage{
mlfdr_fn_tmm(mod.t.stat, prob_mat, param, df, alt.df)
}
\arguments{
\item{mod.t.stat}{numeric of moderated t-statistics.}

\item{prob_mat}{matrix of the conditional probabilities.}

\item{param}{numeric of fitted mu and 1/tau parameters.}

\item{df}{numeric of degrees of freedoms for the null component.}

\item{alt.df}{numeric of degrees of freedoms for the non-null components.}
}
\value{
numeric of mimicked local false discovery rates.
}
\description{
Create mimicked local false discovery rates
}
\examples{
mod.t.cov.df <- create.t.stat.cov.df(exampleY, exampleX, c(1,-1,0,0,0))
fit <- fit.t.model(mod.t.cov.df)
mlfdr <- mlfdr_fn_tmm(mod.t.stat = mod.t.cov.df$mod.t.stat, prob_mat = fit$prob_mat, param = fit$param, df = mod.t.cov.df$df_vec, alt.df = mod.t.cov.df$df_vec)
}
